7.1 Декомпозирует редьюсеры
