+++ src/reducer/name-space.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  DATA: `DATA`,
+  STATE: `STATE`,
+  USER: `USER`,
+};
+
+export default NameSpace;
