+++ src/reducer/state/state.js
@@ -1,12 +1,10 @@
-import {extend} from "./utils.js";
-import films from "./mocks/films.js";
-import {ALL_GENRES} from "./const.js";
+import {extend} from "../utils.js";
+import {ALL_GENRES} from "../../const.js";
 
 const DEFAULT_SHOWED_MOVIES_NUMBER = 8;
 
 const initialState = {
   selectedGenre: ALL_GENRES,
-  films,
   showedMovies: DEFAULT_SHOWED_MOVIES_NUMBER,
   isFullVideoPlayerVisible: false
 };
@@ -15,7 +13,7 @@ const ActionType = {
   CHANGE_GENRE: `CHANGE_GENRE`,
   SHOW_MORE_MOVIES: `SHOW_MORE_MOVIES`,
   RESET_SHOWED_MOVIES_AMOUNT: `RESET_SHOWED_MOVIES_AMOUNT`,
-  CHANGE_VISIBILITY: `CHANGE_VISIBILITY`
+  CHANGE_VISIBILITY: `CHANGE_VISIBILITY`,
 };
 
 const ActionCreator = {
@@ -31,7 +29,7 @@ const ActionCreator = {
     type: ActionType.RESET_SHOWED_MOVIES_AMOUNT,
     payload: null
   }),
-  changeVisibility: () => ({type: ActionType.CHANGE_VISIBILITY})
+  changeVisibility: () => ({type: ActionType.CHANGE_VISIBILITY}),
 };
 
 const reducer = (state = initialState, action) => {
