+++ src/reducer/user/user.js
@@ -1,0 +1,36 @@
+import {extend} from "../utils.js";
+
+const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`
+};
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH
+};
+
+const ActionType = {
+  REQUIRE_AUTHORIZATION: `REQUIRE_AUTHORIZATION`,
+};
+
+const ActionCreator = {
+  requireAuthorization: (status) => {
+    return {
+      type: ActionType.REQUIRE_AUTHORIZATION,
+      payload: status
+    };
+  },
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRE_AUTHORIZATION:
+      return extend({}, state, {
+        authorizationStatus: action.payload
+      });
+  }
+
+  return state;
+};
+
+export {reducer, ActionCreator, ActionType};
